Machine_GPU/deep_ia/requirements.txt
aiofiles==24.1.0
aioice==0.10.0
aiortc==1.11.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.3.0
av==14.3.0
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
contourpy==1.3.1
cryptography==44.0.2
cuda-bindings==12.8.0
cuda-python==12.8.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.13
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fasteners==0.14.1
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
future==0.18.2
google-crc32c==1.7.1
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
humanfriendly==10.0
idna==3.3
ifaddr==0.2.0
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.34.0
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
lap==0.5.12
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
mutagen==1.45.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
onnxslim==0.1.48
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==6.30.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pyds==0.2.0
pyee==13.0.0
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pylibsrtp==0.12.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-json-logger==3.3.0
python-multipart==0.0.20
pytube==15.0.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.3.0
referencing==0.36.2
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
rtsp==1.1.12
scipy==1.15.2
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
systemd-python==234
tensorrt==10.9.0.34
tensorrt_dispatch==10.9.0.34
tensorrt_lean==10.9.0.34
terminado==0.18.1
texttable==1.6.4
tinycss2==1.4.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.91
ultralytics-thop==2.0.14
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
xdg==5
xkit==0.0.0
yt-dlp==2025.3.31
zipp==1.0.0
